requirements.txt
Pillow==5.2.0
boltons==18.0.1
numpy==1.14.5
torch==1.0.1
torchvision==0.2.1
tqdm==4.25.0
tensorboardX==1.4
Cython==0.28.5
matplotlib==2.2.3
click==6.7
